Move registry
